Replace '_' with ' ' in OSM tag hstore query.
A bot on Wikidata changing spaces to underscores in OSM tag values.
There are some cases where we want genuinely want to look for a space in
the OSM tag value. For example "listed_status=Grade II". The
matcher should look for both variants of the OSM tag value, with spaces
and with underscores.

Closes #253