Use wikipedia tag instead of wikipedia + lang
Users have the option to add wikipedia tags in addition to wikidata
tags.

I had chosen to use wikipedia + lang tags like 'wikipedia:en'. I've
since been convinced that it is better to put the language code in the
value.

With this change we now use 'wikipedia' as the tag and for the value we
use: language + ':' + title

Closes: #470