Fix for missing sitelinks
Sometimes we record the first paragraph of a Wikipedia article and later
the sitelink is removed from the Wikidata item. We try to look up the
sitelink and it isn't there.

The code now checks for missing sitelinks, it'll only return extracts
from languages that have a sitelink in the item.

We also invalidate the cache if an extract points to a missing sitelink.